+++ zerver/views/reactions.py
@@ -4,7 +4,7 @@ from django.db import transaction
 from django.http import HttpRequest, HttpResponse
 from django.utils.translation import gettext as _
 
-from zerver.lib.actions import check_add_reaction, do_remove_reaction
+from zerver.actions.reactions import check_add_reaction, do_remove_reaction
 from zerver.lib.emoji import emoji_name_to_emoji_code
 from zerver.lib.exceptions import JsonableError
 from zerver.lib.message import access_message
